+++ src/view/points-list.js
@@ -1,0 +1,7 @@
+const createPoinstListTemplate = () => {
+  return (
+    `<ul class="trip-days"></ul>`
+  );
+};
+
+export {createPoinstListTemplate};
