+++ src/view/summary.js
@@ -1,0 +1,7 @@
+const createSummaryTemplate = () => {
+  return (
+    `<section class="trip-main__trip-info  trip-info"></section>`
+  );
+};
+
+export {createSummaryTemplate};
